add settings layout
