+++ src/components/@settings/PasswordScene/index.jsx
@@ -1,7 +1,6 @@
 import React from 'react'
 import PropTypes from 'prop-types'
 import { withStyles } from '@material-ui/core/styles'
-import Card from '@material-ui/core/Card'
 import CardActions from '@material-ui/core/CardActions'
 import CardContent from '@material-ui/core/CardContent'
 import Button from '@material-ui/core/Button'
@@ -22,8 +21,8 @@ const styles = theme => ({
   },
 })
 
-const PasswordCard = ({ classes }) =>
-  <Card className={classes.card}>
+const PasswordScene = ({ classes }) =>
+  <div className={classes.card}>
     <CardContent>
       <Typography variant="headline" color="primary" className="mb-4">
         Смена пароля
@@ -64,10 +63,10 @@ const PasswordCard = ({ classes }) =>
         Сохранить
       </Button>
     </CardActions>
-  </Card>
+  </div>
 
-PasswordCard.propTypes = {
+PasswordScene.propTypes = {
   classes: PropTypes.object.isRequired,
 }
 
-export default withStyles(styles)(PasswordCard)
+export default withStyles(styles)(PasswordScene)
